@@ -1,7 +1,7 @@
 {
   "name": "content-disposition",
   "description": "Create and parse Content-Disposition header",
-  "version": "0.5.4",
+  "version": "1.0.1",
   "author": "Douglas Christopher Wilson <doug@somethingdoug.com>",
   "license": "MIT",
   "keywords": [
@@ -11,20 +11,19 @@
     "res"
   ],
   "repository": "jshttp/content-disposition",
-  "dependencies": {
-    "safe-buffer": "5.2.1"
+  "funding": {
+    "type": "opencollective",
+    "url": "https://opencollective.com/express"
   },
   "devDependencies": {
-    "deep-equal": "1.0.1",
+    "c8": "^10.1.2",
     "eslint": "7.32.0",
     "eslint-config-standard": "13.0.1",
     "eslint-plugin-import": "2.25.3",
     "eslint-plugin-markdown": "2.2.1",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-promise": "5.2.0",
-    "eslint-plugin-standard": "4.1.0",
-    "istanbul": "0.4.5",
-    "mocha": "9.1.3"
+    "eslint-plugin-standard": "4.1.0"
   },
   "files": [
     "LICENSE",
@@ -33,12 +32,12 @@
     "index.js"
   ],
   "engines": {
-    "node": ">= 0.6"
+    "node": ">=18"
   },
   "scripts": {
     "lint": "eslint .",
-    "test": "mocha --reporter spec --bail --check-leaks test/",
-    "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/"
+    "test": "node --test --test-reporter spec",
+    "test-ci": "c8 --reporter=lcovonly --reporter=text npm test",
+    "test-cov": "c8 --reporter=html --reporter=text npm test"
   }
 }
