@@ -1,7 +1,7 @@
 {
   "name": "mime-types",
   "description": "The ultimate javascript content-type utility.",
-  "version": "2.1.35",
+  "version": "3.0.2",
   "contributors": [
     "Douglas Christopher Wilson <doug@somethingdoug.com>",
     "Jeremiah Senkpiel <fishrock123@rocketmail.com> (https://searchbeam.jit.su)",
@@ -13,27 +13,32 @@
     "types"
   ],
   "repository": "jshttp/mime-types",
+  "funding": {
+    "type": "opencollective",
+    "url": "https://opencollective.com/express"
+  },
   "dependencies": {
-    "mime-db": "1.52.0"
+    "mime-db": "^1.54.0"
   },
   "devDependencies": {
-    "eslint": "7.32.0",
+    "eslint": "8.33.0",
     "eslint-config-standard": "14.1.1",
-    "eslint-plugin-import": "2.25.4",
-    "eslint-plugin-markdown": "2.2.1",
+    "eslint-plugin-import": "2.32.0",
+    "eslint-plugin-markdown": "3.0.1",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "5.2.0",
+    "eslint-plugin-promise": "6.6.0",
     "eslint-plugin-standard": "4.1.0",
-    "mocha": "9.2.2",
+    "mocha": "10.8.2",
     "nyc": "15.1.0"
   },
   "files": [
     "HISTORY.md",
     "LICENSE",
-    "index.js"
+    "index.js",
+    "mimeScore.js"
   ],
   "engines": {
-    "node": ">= 0.6"
+    "node": ">=18"
   },
   "scripts": {
     "lint": "eslint .",
